+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog for NeoFS Contract
 ## [Unreleased]
 
 ### Added
+- Structured containers to Container contract storage (#534)
+- `addStructs`, `createV2` and `getInfo` methods to Container contract (#534)
 
 ### Changed